requirements.txt
alembic
numpy
pandas
pydantic
pydantic-settings
sqlmodel
python-dotenv
openai
langchain-openai
langchain
langchain_core
psycopg2-binary
matplotlib
requests
pydantic_core
